+++ typescript/cli/src/context/context.ts
@@ -61,8 +61,7 @@ export async function contextMiddleware(argv: Record<string, any>) {
 export async function signerMiddleware(argv: Record<string, any>) {
   const { key, requiresKey, multiProvider, strategyPath, chainMetadata } =
     argv.context;
-
-  if (!requiresKey) return argv;
+  if (!requiresKey && !key) return argv;
 
   const strategyConfig = await safeReadChainSubmissionStrategyConfig(
     strategyPath ?? DEFAULT_STRATEGY_CONFIG_PATH,