+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -22,6 +22,7 @@ import { getDomainId } from '../../registry.js';
 import { environment } from './chains.js';
 import { helloWorld } from './helloworld.js';
 import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json';
+import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json';
 import merklyEthAddresses from './misc-artifacts/merkly-eth-addresses.json';
 import merklyNftAddresses from './misc-artifacts/merkly-eth-addresses.json';
 import merklyErc20Addresses from './misc-artifacts/merkly-eth-addresses.json';
@@ -413,6 +414,12 @@ const metricAppContextsGetter = (): MetricAppContext[] => {
       name: 'aave',
       matchingList: senderMatchingList(aaveSenderAddresses),
     },
+    {
+      // https://docs.everclear.org/resources/contracts/mainnet
+      // Messages between HubGateway (Everclear hub) <> EverclearSpoke (all other spoke chains)
+      name: 'everclear_gateway',
+      matchingList: senderMatchingList(everclearSenderAddresses),
+    },
   ];
 };
 
@@ -447,7 +454,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'd834d81-20241125-135658',
+      tag: 'da277d9-20241128-121942',
     },
     gasPaymentEnforcement: gasPaymentEnforcement,
     metricAppContextsGetter,